+++ recipes/configs/dev/qwen3B_async_grpo.yaml
@@ -29,7 +29,7 @@ tokenizer:
 
 # Dataset
 dataset:
-  _component_: torchtune.dev.grpo.gsm8k.gsm8k_dataset
+  _component_: torchtune.dev.rl.gsm8k.gsm8k_dataset
   partition: 1-9/10
 seed: null
 shuffle: False
@@ -105,7 +105,7 @@ training:
     _component_: torchtune.training.lr_schedulers.get_cosine_schedule_with_warmup
     num_warmup_steps: 50
   loss:
-    _component_: torchtune.dev.grpo.loss.GRPOWithChunkedOutputLoss
+    _component_: torchtune.dev.rl.loss.GRPOWithChunkedOutputLoss
     kl_coeff: 0.01
     epsilon: 0.2
